Added New GUI

+++ uninstall.sh
@@ -1,19 +1,19 @@
 #!/bin/bash
 
-#Removing desktop entry
+#Removing All ThemeSaver Files
 rm ~/.local/share/applications/ThemeSaver.desktop
-
-#Removing icon
+sudo rm -r ~/ThemeSaver
 sudo rm /usr/share/icons/ThemeSaver.png
-
-#Removing themesaver file
 sudo rm /usr/local/bin/themesaver
 
 #Removing xfce4-panel-profiles and other dependencies
-sudo apt -y purge xfce4-panel-profiles xdotool python-pil.imagetk python3-pil python3-pil.imagetk fonts-ubuntu
-sudo apt install -f
-sudo apt -y autoremove
-
-#Removing ThemeSaver folder
-sudo rm -r ~/ThemeSaver
-
+if command -v apt &> /dev/null
+then
+    sudo apt -y purge xfce4-panel-profiles xdotool python-pil.imagetk python3-pil python3-pil.imagetk fonts-ubuntu
+    sudo apt install -f
+    sudo apt -y autoremove
+elif command -v pacman &> /dev/null
+then
+    sudo pacman -R xdotool ttf-ubuntu-font-family imagemagick xfce4-panel-profiles
+    fi
+
